requirement.txt
fastapi
uvicorn
pymongo
python-dotenv
sendgrid
